PSO

Description

A short script to do some testing with Particle Swarm Optimization, based on the paper by James Kennedy and Russell Eberhart available at http:www.cs.tufts.educomp150GAhomeworkshw3/_reading6%201995%20particle%20swarming.pdf

Documentation

class Eq a => PSOVect a where

Represents the position and velocity of a particle The four functions are necessary in order to use Points to represent positions and velocities of particles (so we can update a position based on its velocity, update its velocity semi-randomly, etc.)

Methods

pAdd :: a -> a -> a

pScale :: Double -> a -> a

pSubtract :: a -> a -> a

pZero :: a

data PSOCand a

A candidate for a global minimum. Stores both the location of the possible minimum, and the value of the function to minimize at that point.

Constructors

PSOCand

Fields

pt :: a

val :: Double

data Particle a

Particles know their location, velocity, and the best location/value they've visited. Individual particles do not know the global minimum, but the Swarm they belong to does

Constructors

Particle

Fields

pos :: a

position of particle

vel :: a

velocity of particle

pBest :: PSOCand a

best position this particle has found

data Swarm a

A Swarm keeps track of all the particles in the swarm, the function that the swarm seeks to minimize, the parameters, the current iteration (for parameters), and the best location/value found so far

Constructors

Swarm

Fields

parts :: [Particle a]

particles in the swarm

gBest :: PSOCand a

best position found

func :: a -> Double

funtion to minimize

params :: PSOParams

parameters

iteration :: Integer

current iteration

data PSOParams

Holds the parameters used to update particle ; velocities, see Parameter Selection in Particle Swarm Optimization by Yuhui Shi and Russell C. Eberhart

Constructors

PSOParamsStatic Double Double Double

PSOParamsDynamic (Integer -> Double) (Integer -> Double) (Integer -> Double)

randomSwarm

Arguments

:: (PSOVect a, Random a, RandomGen b)

=> b
A random seed

-> Int
Number of particles

-> (a, a)
Bounds to create particles in

-> (a -> Double)
Function to minimize

-> PSOParams
Parameters

-> (Swarm a, b)
(Swarm returned, new seed)

Create a swarm by randomly generating n points within the bounds, making all the particles start at these points with velocity zero.

createSwarm :: PSOVect a => [a] -> (a -> Double) -> PSOParams -> Swarm a

Create a swarm in initial state based on the positions of the particles and the grading function. Initial velocities are all zero.

updateSwarm :: (PSOVect a, RandomGen b) => Swarm a -> b -> (Swarm a, b)

Update the swarm one step, updating every particle's position and velocity, and the best values found so far